import KSAs from ksa.md
Upon importing these KSAs into the file, I realized that some of the enabling objectives in current terminal objectives 1 & 3 don't really fit. I need to create two (or more) new terminal objectives that cover an introduction to the physical parts of the radio as well as a terminal objective that covers changing menu settings with the CHIRP software.